Add a playground to dev, thanks to Fabien Menager
This is code from Fabien Menager's flotr project, which he graciously
has allowed us to borrow.  I have converted most of it to jQuery, and
removed whatever we couldn't use.  There are still tons of flot plugins
and options to be added in, but this is the barebones of it.

Lots of help could be used making it more pretty in different browsers,
making the color wheels work in IE, and general IE testing.