add a better algorithm for highlighting lines
This commit only changes the behaviour of highlighting if lines.show
is true.

This compares the mouse position to the line between any two points,
instead of the current algorithm of just seeing if it is close to
any one point.  The highlights are drawn at half alpha of the series
color and twice the width.  I haven't added any options to
customize it.

TODO item is probably to precompute the lineMagnitude for all the
line segments, as it's currently recomputing them every time the
mouse moves.