make the example highlights a little better.
People copy these examples a lot, so it makes sense to me to keep
them as high quality as possible.  This commit improves the tooltip
to match what is actually highlighted.  Before this patch, if you
had two points that had the same data index and were close to each
other, it would keep the tooltip at the first one you moused over,
even as the highlight moved to the other data point.